chore: polish package.json for npm publishing (#2)
* chore: polish package.json for npm publishing

Add repository, homepage, bugs, author metadata. Include action/ dir
in published files. Pin @types/bun version. Add llm, code-quality,
github-actions keywords for discoverability.


* refactor: consolidate 7 execFile wrappers into shared exec() utility


* refactor: typed ProgressEvent union, consensus uses scanEngine in parallel

Replace stringly-typed progress events with discriminated union.
Refactor runConsensus to delegate to scanEngine and parallelize providers.


* fix: dead re-exports, findingKey collision, Uint16 overflow, misc smells

- Remove dead re-exports from test-gen.ts and fix-gen.ts
- Cache detectTestRunner at module level
- Uint16Array -> Uint32Array in LCS diff
- TOCTOU fix in init.ts
- Extract stripLines() helper in sanitize.ts
- Map lookup in formatSarif
- Parallelize framework detection in test-gen.ts


* fix: use null byte delimiter in findingKey to prevent collisions

Brunt found this in its own code -- colon delimiter can collide
when paths or LLM-generated titles contain colons.


* docs: add CHANGELOG.md for npm publishing


* fix: handle --help flag on subcommands, fix demo root commit

- brunt scan --help now shows help instead of "Unknown flag"
- Demo creates an empty initial commit so HEAD~1 works


---------

+++ src/diff-gen.ts
@@ -95,7 +95,7 @@ function computeLcs(a: string[], b: string[]): Array<[number, number]> {
 
   if (m === 0 || n === 0) return [];
 
-  const full = new Uint16Array((m + 1) * (n + 1));
+  const full = new Uint32Array((m + 1) * (n + 1));
   for (let ii = 1; ii <= m; ii++) {
     for (let jj = 1; jj <= n; jj++) {
       if (a[ii - 1] === b[jj - 1]) {
